Move filter-related functions from filter modal to separate module
